+++ src/app/components/main-pages/pages/item/item.component.scss
