+++ src/app/components/main-pages/pages/item/menu-crud/allergens/allergens.component.scss
