+++ src/app/components/main-pages/pages/item/menu-crud/create-item/create-item.component.scss
