+++ src/app/components/main-pages/pages/item/menu-crud/create-item/create-item.component.ts
@@ -13,9 +13,9 @@ import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 import { MatStepperModule } from '@angular/material/stepper';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { RouterModule } from '@angular/router';
-import { ScheduleEntry } from '../../../../../../../common/menu-categories';
-import { MenuCategoryAvailabilityComponent } from '../../../../../../dialogs/menu-category-availability/menu-category-availability.component';
-import { UploadImageComponent } from '../../../../../../dialogs/upload-image/upload-image.component';
+import { ScheduleEntry } from '../../../../../../common/menu-categories';
+import { MenuCategoryAvailabilityComponent } from '../../../../../dialogs/menu-category-availability/menu-category-availability.component';
+import { UploadImageComponent } from '../../../../../dialogs/upload-image/upload-image.component';
 import { MatDialog } from '@angular/material/dialog';
 import { SafeUrl } from '@angular/platform-browser';
 import { MatDivider } from "@angular/material/divider";
