+++ src/app/components/main-pages/pages/item/menu-crud/properties/properties.component.html
