+++ src/app/components/main-pages/pages/item/menu-crud/recipe/recipe.component.ts
@@ -13,9 +13,9 @@ import { MatSlideToggleModule, MatSlideToggle } from '@angular/material/slide-to
 import { MatStepperModule } from '@angular/material/stepper';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { RouterModule } from '@angular/router';
-import { ScheduleEntry } from '../../../../../../../common/menu-categories';
-import { MenuCategoryAvailabilityComponent } from '../../../../../../dialogs/menu-category-availability/menu-category-availability.component';
-import { UploadImageComponent } from '../../../../../../dialogs/upload-image/upload-image.component';
+import { ScheduleEntry } from '../../../../../../common/menu-categories';
+import { MenuCategoryAvailabilityComponent } from '../../../../../dialogs/menu-category-availability/menu-category-availability.component';
+import { UploadImageComponent } from '../../../../../dialogs/upload-image/upload-image.component';
 import { MatDialog } from '@angular/material/dialog';
 import { SafeUrl } from '@angular/platform-browser';
 import { MatDivider } from "@angular/material/divider";
@@ -25,11 +25,10 @@ import { BooleanInput } from '@angular/cdk/coercion';
 import { MatCheckbox } from "@angular/material/checkbox";
 import { MatAccordion, MatExpansionPanel, MatExpansionPanelHeader, MatExpansionPanelTitle } from '@angular/material/expansion';
 
-import { MatChip } from "../../../../../../../../../node_modules/@angular/material/chips/index";
-
+import { MatChip } from "../../../../../../../../node_modules/@angular/material/chips/index";
 
 @Component({
-  selector: 'app-properties',
+  selector: 'app-recipe',
    standalone: true,
   imports: [
     CommonModule,
@@ -47,41 +46,21 @@ import { MatChip } from "../../../../../../../../../node_modules/@angular/materi
     MatTooltipModule,
     MatChipsModule,
     MatDatepickerModule,
-    MatNativeDateModule,
-    MatAccordion,
-    MatExpansionPanel,
-    MatExpansionPanelHeader,
-    MatExpansionPanelTitle
+    MatNativeDateModule
+
 ],
 providers: [  
   MatDatepickerModule,  
 ],
-
-  templateUrl: './properties.component.html',
-  styleUrl: './properties.component.scss'
+  templateUrl: './recipe.component.html',
+  styleUrl: './recipe.component.scss'
 })
-export class PropertiesComponent  implements OnInit {
-  
-  accordion = viewChild.required(MatAccordion);
-
-  confirmCreate(): void {
-
-  }
-  fb: any;
-
-
-
+export class RecipeComponent implements OnInit{
+confirmCreate() {
+throw new Error('Method not implemented.');
+}
   ngOnInit(): void {
-
-    this.propertiesForm = this.fb.group({
-  
-    });
+    throw new Error('Method not implemented.');
   }
-
-
- propertiesForm!: FormGroup;
-
-
-
 
 }
